refactor(test): extract shared createTempFile/cleanupTempFile helpers
Addresses review comment by @dinohamzic.
