+++ packages/cli/src/commands/cat.test.ts
@@ -1,5 +1,3 @@
-import fs from 'node:fs/promises'
-import os from 'node:os'
 import { join, resolve } from 'node:path'
 import { stripVTControlCharacters } from 'node:util'
 import { Command } from 'commander'
@@ -7,27 +5,12 @@ import { afterEach, beforeEach, describe, expect, it, type Mock, vi } from 'vite
 import { ExitCode } from '../exit-codes'
 import { resetOutputConfig } from '../output'
 import { type CatOptions, createCatAction } from './cat'
+import { cleanupTempFile, createTempFile } from './test-helpers'
 
 // Test file paths relative to project root (tests are run from root)
 const HELLO_WORLD_FILE = join('examples', '1_hello_world.deepnote')
 const BLOCKS_FILE = join('examples', '2_blocks.deepnote')
 const INTEGRATIONS_FILE = join('examples', '3_integrations.deepnote')
-
-async function createTempFile(content: string): Promise<string> {
-  const tempDir = await fs.mkdtemp(join(os.tmpdir(), 'deepnote-cat-test-'))
-  const filePath = join(tempDir, 'test.deepnote')
-  await fs.writeFile(filePath, content, 'utf8')
-  return filePath
-}
-
-async function cleanupTempFile(filePath: string): Promise<void> {
-  try {
-    await fs.unlink(filePath)
-    await fs.rmdir(join(filePath, '..'))
-  } catch {
-    // Ignore cleanup errors
-  }
-}
 
 /** Default options for testing */
 const DEFAULT_OPTIONS: CatOptions = {}
